requirements.txt
onnxruntime
rembg
pillow
tk